+++ components/ChatSelector.vue
@@ -48,20 +48,11 @@ input {
     flex-direction: column;
 }
 
-@media only screen and (max-width:1350px) {
-    .container {
-        width: 20rem;
-    }
-}
 
 
 @media only screen and (max-width:1120px) {
-    .container {
-        width: 18rem;
-    }
-
     input {
-        width: 15.75rem;
+        width: 100%;
     }
 }
 </style>
